refactor: extract hook logic to single file
